Restructure system pipeline, fix initial render bug
I decided to merge the different stages together, and group most systems
under the Update stage. Update only runs when not waiting for input.
This reduces code repetition in the pipeline creation.

I fixed the bug where 50% of the time the screen would start pitch
black. The system that updated viewsheds and the system that applies
the player viewshed to the map ran concurrently, meaning half the time
the player viewshed was applied before being updated.